chore: the logic of the form is placed in a separate component
