+++ my-vue-app/src/App.vue
@@ -1,7 +1,7 @@
 <script setup>
 import Header from './components/Header/Header.vue';
 import MainBanner from './components/MainBanner/MainBanner.vue';
-import Footer from './components/Footer.vue';
+import Footer from './components/Footer/Footer.vue';
 </script>
 
 <template>
