--- my-vue-app/src/assets/styles/components/footer.scss
@@ -1,190 +1,0 @@
-@import './main/index.scss';
-
-.footer {
-  $self: &;
-
-  min-height: 844px;
-  padding: 4rem 2rem;
-  text-align: left;
-  background-color: $bg-footer;
-
-  @media (max-width: $breakpoint-xs) {
-    position: relative;
-    top: auto;
-    width: 100%;
-    height: auto;
-    padding: 2rem 1rem;
-  }
-
-  &__container {
-    max-width: 1340px;
-    margin: 0 auto;
-  }
-
-  &__content {
-    display: flex;
-
-    @media (max-width: $breakpoint-xl) {
-      display: grid;
-      grid-template-columns: 1fr;
-      gap: 3rem;
-    }
-
-    @media (max-width: $breakpoint-md) {
-      flex-direction: column;
-      gap: 2rem;
-    }
-  }
-
-  &__info {
-    flex: 1;
-    display: flex;
-    flex-direction: column;
-    justify-content: space-between;
-
-    @media (max-width: $breakpoint-xl) {
-      display: grid;
-      grid-template-columns: 1fr;
-      gap: 2rem;
-      justify-content: stretch;
-    }
-  }
-
-  &__links {
-    display: flex;
-    flex-direction: column;
-    font-size: 17px;
-    gap: 20px;
-
-    @media (max-width: $breakpoint-xl) {
-      display: grid;
-      grid-template-columns: repeat(auto-fit, minmax(150px, 1fr));
-      gap: 1.5rem;
-    }
-
-    @media (max-width: $breakpoint-md) {
-      grid-template-columns: 1fr;
-    }
-  }
-
-  &__form {
-    max-width: 614px;
-    padding: 0 10rem;
-
-    @media (max-width: $breakpoint-xl) {
-      max-width: none;
-      padding: 0;
-    }
-
-    @media (max-width: $breakpoint-md) {
-      max-width: 100%;
-      width: 100%;
-    }
-
-    &-title {
-      font-family: $font-family-base;
-      font-weight: 600;
-      font-size: 28px;
-      color: $bg-primary;
-    }
-
-    &-description {
-      font-family: $font-family-base;
-      font-weight: 400;
-      font-size: 16px;
-      line-height: 100%;
-      color: $bg-primary;
-
-      @media (max-width: $breakpoint-md) {
-        font-size: 1.2rem;
-        line-height: 1.4;
-      }
-    }
-
-    &-group {
-      margin-bottom: 1.7rem;
-      color: $color-form-placeholder;
-    }
-
-    &-input {
-      width: 472px;
-      padding: 32px 0;
-      font-size: 1rem;
-      color: $color-form-placeholder;
-      background: transparent;
-      border: none;
-      border-bottom: 1px solid $border-form;
-
-      @media (max-width: $breakpoint-xl) {
-        width: 100%;
-        max-width: none;
-      }
-
-      &:hover {
-        border-bottom-color: $ui-button-blue;
-      }
-
-      &:focus,
-      &:not(:placeholder-shown) {
-        color: $text-primary;
-        border-bottom-color: $ui-button-blue;
-      }
-    }
-
-    &-textarea {
-      width: 472px;
-      min-height: 100px;
-      font-size: 1rem;
-      color: $color-form-placeholder;
-      background: transparent;
-      border: 1px solid $border-form;
-      border-radius: 4px;
-
-      @media (max-width: $breakpoint-xl) {
-        width: 100%;
-        max-width: none;
-      }
-
-      &:focus,
-      &:not(:placeholder-shown) {
-        color: $text-primary;
-        border-bottom-color: $ui-button-blue;
-      }
-    }
-
-    &-select {
-      cursor: pointer;
-      appearance: none;
-
-      &:focus, 
-      &:not(:placeholder-shown) {
-        color: $color-form-placeholder;
-      }
-
-      &-icon {
-        position: absolute;
-        width: 10px;
-        height: 6px;
-        pointer-events: none;
-        transform: translate(-82%, 700%);
-      }
-    }
-
-    &-button {
-      display: flex;
-      justify-content: space-between;
-      width: 100%;
-      max-width: 480px;
-
-      @media (max-width: $breakpoint-xl) {
-        max-width: none;
-      }
-
-      @media (max-width: $breakpoint-xs) {
-        max-width: 100%;
-        width: 100%;
-        flex-direction: column;
-      }
-    }
-  }
-}
