+++ my-vue-app/src/assets/styles/components/footer/footer.scss
@@ -1,0 +1,69 @@
+@import '../main/index.scss';
+
+.footer {
+  $self: &;
+
+  min-height: 844px;
+  padding: 4rem 2rem;
+  text-align: left;
+  background-color: $bg-footer;
+
+  @media (max-width: $breakpoint-xs) {
+    position: relative;
+    top: auto;
+    width: 100%;
+    height: auto;
+    padding: 2rem 1rem;
+  }
+
+  &__container {
+    max-width: 1340px;
+    margin: 0 auto;
+  }
+
+  &__content {
+    display: flex;
+
+    @media (max-width: $breakpoint-xl) {
+      display: grid;
+      grid-template-columns: 1fr;
+      gap: 3rem;
+    }
+
+    @media (max-width: $breakpoint-md) {
+      flex-direction: column;
+      gap: 2rem;
+    }
+  }
+
+  &__info {
+    flex: 1;
+    display: flex;
+    flex-direction: column;
+    justify-content: space-between;
+
+    @media (max-width: $breakpoint-xl) {
+      display: grid;
+      grid-template-columns: 1fr;
+      gap: 2rem;
+      justify-content: stretch;
+    }
+  }
+
+  &__links {
+    display: flex;
+    flex-direction: column;
+    font-size: 17px;
+    gap: 20px;
+
+    @media (max-width: $breakpoint-xl) {
+      display: grid;
+      grid-template-columns: repeat(auto-fit, minmax(150px, 1fr));
+      gap: 1.5rem;
+    }
+
+    @media (max-width: $breakpoint-md) {
+      grid-template-columns: 1fr;
+    }
+  }
+}
