+++ my-vue-app/src/assets/styles/components/header/header.scss
@@ -23,8 +23,6 @@
     font-family: $font-family-base;
   }
 
-
-
   &__nav {
     display: grid;
     align-items: center;
